src/mcp-ecosystem/servers/pyproject.toml
[project]
name = "mem0-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "mem0ai>=0.1.55",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.2.1",
    "starlette>=0.46.0",
]
